+++ lang/cs/role-list.php
@@ -1,0 +1,15 @@
+<?php
+
+return [
+
+    'close' => 'Zavřít',
+    'no-additional-permissions' => 'Žádná další oprávnění.',
+
+    'UserPolicy@crud' => 'Spravovat svůj profil',
+    'UserPolicy@viewAny' => 'Zobrazit ostatní uživatele',
+    'UserPolicy@viewRoleList' => 'Zobrazit role a oprávnění',
+    'UserPolicy@administrate' => 'Spravovat ostatní uživatele',
+    'UserPolicy@setRole' => 'Měnit uživatelské role',
+    'UserPolicy@merge' => 'Slučovat uživatelské účty',
+
+];