+++ lang/da/role-list.php
@@ -1,0 +1,15 @@
+<?php
+
+return [
+
+    'close' => 'Luk',
+    'no-additional-permissions' => 'Ingen yderligere rettigheder.',
+
+    'UserPolicy@crud' => 'Administrere egen profil',
+    'UserPolicy@viewAny' => 'Se andre brugere',
+    'UserPolicy@viewRoleList' => 'Se roller og rettigheder',
+    'UserPolicy@administrate' => 'Administrere andre brugere',
+    'UserPolicy@setRole' => 'Ændre brugerroller',
+    'UserPolicy@merge' => 'Sammenlægge brugerkonti',
+
+];