+++ lang/en/role-list.php
@@ -1,0 +1,15 @@
+<?php
+
+return [
+
+    'close' => 'Close',
+    'no-additional-permissions' => 'No additional permissions.',
+
+    'UserPolicy@crud' => 'Manage own profile',
+    'UserPolicy@viewAny' => 'View other users',
+    'UserPolicy@viewRoleList' => 'View roles and permissions',
+    'UserPolicy@administrate' => 'Manage other users',
+    'UserPolicy@setRole' => 'Change user roles',
+    'UserPolicy@merge' => 'Merge user accounts',
+
+];